Remove build files

+++ src/sevensegment.c
@@ -34,7 +34,7 @@ uint8_t ssDigit(uint8_t dec) {
 
 
 
-void port(uint8_t digit, int portarray[]){
+void port(uint8_t digit, uint8_t portarray[]){
     int i;
     for (i = 0; i < 7; i++){
         if (digit&(1<<i)){
@@ -49,8 +49,8 @@ void ssDisplay(int value){
     uint8_t digit0;
     uint8_t digit1;
     
-    int port0array[7] = {PC0, PC1, PC4, PC5, PC6, PC7, PB7};
-    int port1array[7] = {PB0, PB1, PB2, PB3, PB4, PB5, PB6};
+    uint8_t port0array[7] = {PC0, PC1, PC4, PC5, PC6, PC7, PB7};
+    uint8_t port1array[7] = {PB0, PB1, PB2, PB3, PB4, PB5, PB6};
 
     /* Cap value at 99 */
     if (value > 99) {
